pyproject.toml
[tool.poetry]
name = "gifsync"
version = "0.0.0"
description = "Syncs gifs and audio"
authors = ["dhsavell <dhsavell05@gmail.com>"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "^3.7"
pillow = "^6.2"
wand = "^0.5.8"
click = "^7.0"
numpy = "^1.18"
scipy = "^1.4"
librosa = "^0.7.1"
pysndfx = "^0.3.6"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pylint = "^2.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
